Fix docker push & combine workflows
Combining these means that the PR build will always work the same
as the main build. The fix for the push issue turned out to be that
the name of the image (specified by the tags) needed to be prefixed
by the name of the repo (i.e: ghcr.io).